+++ src/com/intellij/idea/plugin/hybris/impex/settings/ImpexSettings.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -23,11 +23,18 @@ import com.intellij.openapi.components.BaseState
 data class ImpexSettings(
     var folding: ImpexFoldingSettings = ImpexFoldingSettings(),
     var completion: ImpexCompletionSettings = ImpexCompletionSettings(),
+    var documentation: ImpexDocumentationSettings = ImpexDocumentationSettings(),
 )
 
 data class ImpexFoldingSettings(
     var enabled: Boolean = true,
     var useSmartFolding: Boolean = true,
+)
+
+data class ImpexDocumentationSettings(
+    var enabled: Boolean = true,
+    var showTypeDocumentation: Boolean = true,
+    var showModifierDocumentation: Boolean = true,
 )
 
 class ImpexCompletionSettings : BaseState() {